refactor(commander): split commanderMetrics into per-section helpers
